+++ integration-tests/emulated/tests/people/people-polkadot/src/tests/mod.rs
@@ -13,5 +13,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod governance;
 mod reap_identity;
 mod teleport;